[REFACTOR] Refactor File Position to Inhancemnet Managements
