+++ app/src/main/AndroidManifest.xml
@@ -52,7 +52,7 @@
             android:exported="true">
 
         </activity>
-        <activity android:name=".presentation.screen.order.SandwitchActivity"
+        <activity android:name=".presentation.screen.order.sandwitch.SandwitchActivity"
             android:exported="true">
 
         </activity>
@@ -86,12 +86,12 @@
                     android:scheme="kakao432c8e499713dc07778219f7dc6ab3b1" />
             </intent-filter>
         </activity>
-        <activity android:name="com.example.pocky.presentation.screen.order.SmilesuppActivity" />
-        <activity android:name="com.example.pocky.presentation.screen.order.BreakfastActivity" />
+        <activity android:name="com.example.pocky.presentation.screen.order.smilesupp.SmilesuppActivity" />
+        <activity android:name="com.example.pocky.presentation.screen.order.breakfast.BreakfastActivity" />
         <activity android:name="com.example.pocky.presentation.screen.order.common.ConfirmationActivity" />
-        <activity android:name="com.example.pocky.presentation.screen.order.GroupmenuActivity" />
-        <activity android:name="com.example.pocky.presentation.screen.order.SaladActivity" />
-        <activity android:name="com.example.pocky.presentation.screen.order.WrappActivity" />
+        <activity android:name="com.example.pocky.presentation.screen.order.groupmenu.GroupmenuActivity" />
+        <activity android:name="com.example.pocky.presentation.screen.order.salad.SaladActivity" />
+        <activity android:name="com.example.pocky.presentation.screen.order.wrapp.WrappActivity" />
     </application>
 
 </manifest>
