+++ app/src/main/java/com/example/pocky/presentation/screen/order/breakfast/BreakfastActivity.java
@@ -1,13 +1,14 @@
-package com.example.pocky.presentation.screen.order;
+package com.example.pocky.presentation.screen.order.breakfast;
 
 import android.os.Bundle;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.example.pocky.R;
@@ -15,31 +16,21 @@ import com.example.pocky.R;
 import java.util.Arrays;
 import java.util.List;
 
-public class SmilesuppActivity extends AppCompatActivity {
+public class BreakfastActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_smilesupp);
+        setContentView(R.layout.activity_breakfast);
 
         RecyclerView recyclerView = findViewById(R.id.recyclerView);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
         List<Integer> imageList = Arrays.asList(
-                R.drawable.resize_baconcheesewedgepotato,
-                R.drawable.resize_cheesewedgepotato,
-                R.drawable.resize_chickenbaconminiwrap,
-                R.drawable.resize_chips,
-                R.drawable.resize_chocolatechip,
-                R.drawable.resize_cornsoup,
-                R.drawable.resize_doublechocolatechip,
-                R.drawable.resize_hashbrown,
-                R.drawable.resize_milk,
-                R.drawable.resize_mushroomsoup,
-                R.drawable.resize_oatmealrasin,
-                R.drawable.resize_raspberrycheesecake,
-                R.drawable.resize_wedgepotato,
-                R.drawable.resize_whitechocomacadamia
+                R.drawable.resize_hameggcheese,
+                R.drawable.resize_hameggcheesewrap,
+                R.drawable.resize_westerneggcheese,
+                R.drawable.resize_westerneggcheesewrap
         );
 
         BreadAdapter adapter = new BreadAdapter(imageList);
