+++ app/src/main/java/com/example/pocky/presentation/screen/order/sandwitch/SandwitchActivity.java
@@ -1,4 +1,4 @@
-package com.example.pocky.presentation.screen.order;
+package com.example.pocky.presentation.screen.order.sandwitch;
 
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
